requirements.txt
absl-py==0.9.0
appdirs==1.4.3
asn1crypto==1.3.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
configparser==5.0.0
cryptography==2.8
cycler==0.10.0
cytoolz==0.10.1
dask==2.14.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
docker-pycreds==0.4.0
dparse==0.5.0
entrypoints==0.3
filelock==3.0.12
gast==0.2.2
gitdb==4.0.2
GitPython==3.1.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pluggy==0.13.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
safety==1.8.7
scikit-image==0.16.2
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.1
subprocess32==3.5.4
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.5.4
wandb==0.8.31
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==2.2.0
